memcache updates
git-svn-id: svn://svn.caucho.com/resin/trunk@8479 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/server/distcache/CacheData.java
@@ -46,7 +46,7 @@ public final class CacheData extends MnodeValue {
                    long valueLength,
                    long version,
                    HashKey cacheKey,
-                   int flags,
+                   long flags,
                    long accessTime,
                    long expireTimeout,
                    long idleTimeout)
